SPAZA-297 Allow fixed parameters during fitting
Allow pointing model parameters to be fixed to their old values by
introducing a `keep_disabled_params` parameter to `PointingModel.fit()`.
If True, the old model values are kept and subtracted from the data
before fitting the enabled parameters. These residual fits are then
added back to the original parameters to arrive at the final parameters.

If False (the default for now and the original behaviour), all model
parameters are first zeroed before fitting the enabled parameters.
This behaviour is deprecated and `keep_disabled_params` will be forced
to True and removed in a future release. The original behaviour can be
achieved by explicitly zeroing the model before calling `fit()`. Warn
the user about this.

Clean up `fit()` by not abusing `self` to generate columns in the
design matrix. Improve the docstring for `PointingModel` by pointing
out that it can be initialised from an existing model. Add unit tests
to check treatment of enabled and disabled parameters.

This addresses SARAO ticket SPR1-407.

[v1: redid the tests in pytest]