SPAZA-267 Check that temperature is set with RH
The default temperature of 0 degrees Celsius is problematic, since the
relative humidity may be set via kwarg while the temperature is left
unspecified. The resultant delay will then most likely be wrong, and
we would be none the wiser.

Rather raise a ValueError if the relative humidity has some positive
(i.e. non-zero) value and the temperature has not been set. Detect the
latter by changing the default temperature to a non-physical value,
since `u.quantity_input` does not play nicely with `None`.