SPAZA-266 New angle_to_string function
This converts an Astropy Angle object to a string (or an array of
strings if multidimensional), with a focus on high precision and
backwards compatibility with older katpoint.

The chosen precision preserves the angle subtended by a micron at a
distance equal to the Earth's diameter. This was derived from the
accuracy required for Antenna geodetic coordinates (see SPAZA-260),
but should also suffice for sky coordinates.

The numerical representation is either decimal or sexagesimal, like
in katpoint v0.x, but the string now also has a suffix indicating the
unit ('d' for degree or 'h' for hour angle). This will allow the string
to initialise an Astropy Angle directly without the need for `to_angle`.
Older katpoint will be patched to interpret this suffix as well.

Take note that 'dms' and 'hms' separators are not supported, and
neither is a unit of radians.

Add several unit tests to check corner cases as well as round-tripping.