SPAZA-266 Introduce Target.from_description
Move the `construct_target_params` function to a class method called
`Target.from_description`. This now constructs a complete Target object
and not just its component parts.

Although the original function's name does not start with an underscore,
it was never meant to be part of the public API, so no deprecation
warning is necessary. Just cull it unceremoniously.