SPAZA-266 Rework Target initialisation
This follows the lead of SPAZA-260 (!10), which did this for Antennas.

Allow additional parameters to override the contents of a description
string or existing Target object during initialisation. Previously
the additional parameters were only used when the first parameter was
a Body. The new scheme seems more natural.

There is still work to be done... One cannot override the name on the
one hand, and one should not be able to override the body type (first
tag) on the other hand.