SPAZA-266 Move body type from Target to Body
The body type aka first tag is special, because it is associated with
the Body inside the Target. It is therefore a bad idea to expose it
in editable form on the Target. Split `Target.tags` into two parts:

  - The body type (and possible subtype) moves to `Body.tag`.
  - The rest of the tags are stored in the new `Target.user_tags`.

The `user_tags` list is fully user-editable and is also passed into the
Target initialiser. The `tags` attribute is now a read-only property
that concatenates the body and user tags.

The `Body.tag` is initialised on construction of the Body object. It is
purely up to the Body to figure out its tag, and Target is not involved
with this anymore. It is generally a proxy for the relevant Body class,
but there are exceptions:

  - The FixedBody can be 'radec', 'gal' or 'xephem radec', depending
    on the frame and the way it was constructed.
  - The EarthSatelliteBody can be 'tle' or 'xephem tle'. Unfortunately
    the EDB-based version cannot yet be saved to TLE, otherwise we
    could possibly have done away with the 'xephem' body type.
  - What about 'B1900' and 'B1950' subtypes? They affect construction
    of the FixedBody...

A side effect is that it is now finally possible to construct a Target
from just a Body object; previously, tags had to be provided too.

Add unit tests to check that the various tags are as expected. Add
another Earth satellite to the unit tests, both in EDB and TLE form,
which will eventually be used to check conversion between the formats.