SPAZA-291 Remove star database and body type
        Stars have their moment and then they die. -- Nick Cave

These are warts from the early days of katpoint. The stars database was
copied from PyEphem, simply because PyEphem had it. It was briefly
useful for exploring optical pointing of early dishes, but it could just
as well live in a separate catalogue. Make good on this suggestion and
move the database to an EDB file in the scripts directory (still to be
finessed into a proper database).

There is no need to assign stars a body type and special support in
`Catalogue`. In fact, the `star` body type is what hastened its demise,
because it does not play well with the changes to tags planned for
SPAZA-266. Loading the STARS database into Bodies also adds 100 ms to
katpoint's startup for no benefit.

Crash out if a user attempts to set `add_stars=True` in `Catalogue`,
but politely warn if `add_stars=False` instead.