SPAZA-291 Deprecate add_specials in Catalogue
This is another wart. It is rarely a good idea to add all Solar System
bodies to a Catalogue, and they are also relatively easy to add
manually. As with add_stars, keep the parameter and raise ValueError if
the user sets it to True, but warn if it is set to False. The latter
still occasionally happens for some reason, probably because the
default value for add_specials was True before 2010 (see 6087d2c) and
some of us are creatures of habit...

Redo the unit tests with an explicit list of targets.