SPAZA-266 Figure out default name from TLE
We can do better than 'Unnamed Satellite'... :joy: Also, leave the
default name to the Body and not the Target. Picked the NORAD ID
above the COSPAR ID because it is easier to construct. The EDB version
of EarthSatelliteBody does not have this issue, since the name is
embedded in the EDB.