SPAZA-266 Clarify Target/Body name handling
Make Bodies intrinsically nameless, just like `SkyCoord`. An exception
is any body with a "special" tag, which only cares about the name.

Each Body now offers a default name derived from the coordinates or
properties of the body (i.e. names like "Ra: xxx Dec: xxx"). This is a
useful fallback when the Target does not have an explicit name, to avoid
upsetting Catalogues, archive metadata extractors and more with missing
names. This scheme also applies to Earth satellites, with default names
based on a circular orbit approximation with three parameters (two
orientation angles and one orbit size).

The `Body.default_name` is a property and hence read-only. The same is
done for the `Body.tag` to avoid inadvertent changes. The tag is now
deduced from the body properties instead of getting passed around.
The base class therefore does not need an initialiser anymore.

The `FixedBody` and `EarthSatelliteBody` ignore the names in the EDB
format (at the Body level). The former also reverts to a normal 'radec'
tag, while the latter remains 'xephem tle' due to the differences in
format. EDB is much more rudimentary than TLE, storing only the most
important orbital elements and no satellite IDs, so we still need to
save it as EDB.

The Target now handles any preferred names and aliases, only reverting
to the default Body name if the preferred name is empty. This makes it
easier to detect unnamed Targets, too. Properties keep both the main
and alternate names read-only to prevent strangeness in e.g. Catalogues.