SPAZA-266 Support flux models for all body types
Flux models were not supported for *special* and *xephem* bodies because
these bodies have 0 and 1 location fields, respectively, while the rest
have 2 location fields and the flux model was actually hardwired to the
fifth field. This is inconsistent and also complicates description
string creation.

The solution is to pop off fields from the start of the list as the
names, tags and locations are processed. This leaves the optional flux
model as the next remaining field at the end of description string
parsing. Creating the description string is also simpler, since the
flux model can be tacked on the end of any body.

Add a unit test to check that flux models can be successfully added to
any body type.