SPAZA-293 flux_freq_MHz, flux_density -> Quantity
This changes the `flux_freq_MHz` variable with its hardcoded unit of
MHz to an Astropy Quantity. This is passed as parameter to Target and
Catalogue, and ultimately ends up in `FluxDensityModel.flux_density`.
Let the `flux_density` function return a Quantity with a unit of Jy,
while we are at it, and let the flux_limit parameter of Catalogue
accept a Jy Quantity.

Update docstring examples and check the unit when setting the
parameters via the `u.quantity_input` decorator.