SPAZA-293 Rework FluxDensityModel construction
Turn `min_freq_MHz` and `max_freq_MHz` into Quantities, losing the
hardcoded units. Split conversion to and from description strings
to separate methods so that the standard initialiser is not
overloaded anymore. This allows the use of the `quantity_input`
decorator for unit checking, which can only tolerate `None` for the
first parameter but not a description string. In other words, replace

  FluxDensityModel('100 200 1')

with

  FluxDensityModel.from_description('100 200 1')

in all places. The __str__ representation is simplified to just the
description string (who needs three different string representations?).
Improve error messages for description string parser, and keep the
slightly icky `FluxError` for now.