SPAZA-293 Clean up Catalogue units
Replace all '*_deg' variables that used to be angles in degrees with
Quantities. Rework the angle bounds checks in `Catalogue.iterfilter`
to use `Angle.is_within_bounds`, which simplifies the if-statements
a lot. Wrap azimuth angles using `Longitude`. Let `closest_to` return
an `Angle` and not a float in degrees. Check the validity of angle
limits and add unit tests for the checks as well.

Get rid of duplicate calls to Target.azel(), which probably improved
the unit test times by 10%. Try to keep variables as Quantities for
longer before casting it to specific units.

Update all the docstring examples to use units. Use the same trick as
in `DelayCorrection` to share the Parameters section of the docstring
between `iterfilter` and `filter`.