SPAZA-293 Let baselines and UVW have units
Introduce Astropy units to:

  - `Antenna.baseline_toward` (CartesianRepresentation, in metres)
  - `Target.geometric_delay` (Quantities, in seconds and dimensionless)
  - `Target.uvw` (Quantity, in metres)

Tweak the tests to compare against numbers with units, which are more
interpretable. Introduce `np.array_equal` in `baseline_toward`, which
actually surprisingly allows the calculation of multiple baselines
(unadvertised for now). Don't interpret the `delay_rate` Quantity as a
bool as this is deprecated - compare it against 0 instead.