SPAZA-266 Restore EDB names in description string
When writing out an "xephem" Target (currently only TLEs), stuff the
name(s) back into the EDB string if it's not the Body default name.
This restores the EDB string the way it was before constructing the
Target.

Add a unit test that checks that the description string round-trips in
this case (starting and ending with no name field).