SPAZA-266 Fix rogue "xephem" tag
Since the Body is now the sole arbiter of the body type tag, pop it
while parsing the description string. This ensures that an
"xephem radec" Target is transformed to an "radec" Target without
"xephem" tagging along. It is also no longer necessary to run
`.lower()` on the body type tag, because it is guaranteed to be
lower-case by the Body.