SPAZA-266 Make aliases a tuple, add names property
To keep the aliases immutable, turn it into a tuple in __init__ since
a tuple is the ideal immutable sequence. This will also warn the user on
attempts to modify the attribute, contrary to a list.

Add a convenience property that aggregates all the Target names into
a tuple. This should replace code like `[target.name] + target.aliases`,
which is popular (even in user scripts) and will break now that
`aliases` is a tuple.