SPAZA-296 Fix sdist deficiency in tox
The sdist step in tox does not include the same files as just running
`python setup.py sdist` (probably distutils vs setuptools or similar).
Not having NEWS.rst causes the "sdist-make" step to fail when running
the tox py38 command. Add these bits to MANIFEST.in to hopefully ensure
their inclusion.

While we are at it, remove some packages from gitlab-ci-requirements.txt
that are there because of other packages in the list.