SPAZA-287 Calculate geometric delays in ITRS frame
The original katpoint approach worked in the ENU frame relative to the
reference location, with ENU antenna locations and a direct conversion
of the target (az, el) to an ENU unit vector.

The katpoint_vs_calc study in 2020 revealed that ITRS (or XYZ) baselines
are more accurate when compared to Calc, at least when doing ENU with a
mix of homegrown routines and Astropy. The difference is on the order
of 16 ps for MeerKAT (about 0.15" position offset) and seems to be
linked to diurnal aberration between topocentric and geocentric frames.
See e.g. 63bd909.

Fix this as part of the push to improve MeerKAT astrometry. Convert
the target (az, el) to an ITRS unit vector while ensuring that the
distance towards the target is ignored. Also calculate the antenna
baselines towards the reference location in XYZ. Then take the dot
product as before.