SPAZA-287 Fix basic (az, el) delay tests
With ENU-based delays it was possible to define an (az, el) target
that is perfectly perpendicular to a given baseline, and it remained
that way forever. This allowed for some simple tests.

Now that the delays are done in ITRS, this has been disrupted. There
is definitely some diurnal aberration mixed into the translation
between AltAz and ITRS, since we can minimise the error by placing the
target due West (or East) of the array.

Use a Western target and relax the exact tests.