SPAZA-287 Fix Calc comparison tests
The first place we encountered this ENU vs ITRS discrepancy was in the
katpoint_vs_calc study, so it is natural to update the tests based on
that study as well.

There is no more need to track the minimum error because we have closed
the gap with the Astropy ITRS implementation in the study. The minimum
error is now around 0.1 ps. The remaining residuals range from about
6 ps on MeerKAT to 19 ps on MeerKAT+.

The latter seems worse than the 12-16 ps obtained with ENU baselines in
version 1.0a1, until you realise that s0121 was the worst antenna in
the original study but not for 1.0a1. The maximum discrepancy between
katpoint 1.0a1 and Calc was 39 ps on s0105, so add that antenna to the
tests to ensure that the error does not sneak back.

Add a reminder to extend the Calc tests to the Sun/Moon and TLEs, to
verify that our act of stripping of AltAz distances is correct.