SPAZA-287 Add GCRS and ENU baseline options
I started with topocentric ENU baselines. I then tried out geocentric
ITRS baselines, followed by a full implementation of the VLBI consensus
model (geocentric GCRS baselines). While the latter matches Calc very
well, it does not fix the astrometry issues in the 3C 286 datasets.

I then tried out GCRS baselines and made it topocentric by specifying
`obsgeoloc` and `obsgeovel` parameters (out of habit). That improved
the astrometry, and the discrepancy with Calc reminded me of the ENU
errors... So back we went to ENU and it is the same!

It turns out that the topocentric part is the key, which gets rid of
the diurnal aberration residual mentioned before. This was also verified
by turning the geocentric consensus model into a topocentric version
by replacing all occurrences of `VE` (the barycentric velocity of the
geocenter) with `V1` (the barycentric velocity of the reference
location) in `_vlbi_delays` and getting similar results.