SPAZA-287 Revert to 1.0a1 ENU baselines for now
This is to check that v1.0a1 can correct the 3C286 offsets via katdal's
KRETRACK mechanism. It gives very similar results to the topocentric
GCRS baselines, while also supporting (az, el) and offset targets.

In the long run, we still want to replace ENU baselines with GCRS to
avoid the dependence on ENU offsets in the delay model. The geometric
delay should only depend on the antenna's `EarthLocation`. We also want
to restore the Calc unit tests by comparing it against a geocentric
version of `DelayCorrection`.