SPAZA-295 Ignore some TimeDeltaMissingUnitWarnings
These corner cases happen when an operator goes to Astropy first:

- TimeDelta + Timestamp -> TimeDelta.__add__(time_delta, ts)
- Time - Timestamp -> Time.__sub__(time, ts)

In both cases the katpoint.Timestamp object `ts` is cast to a TimeDelta
without a format specifier, triggering the warning in Astropy>=5.1.
It then proceeds to raise NotImplemented, which passes the baton to
Timestamp.__radd__ and Timestamp.__rsub__, respectively, which does
the right thing.

Once the deprecation warning becomes a proper error (Astropy 6.0?)
we can remove this workaround, as the code will just hit NotImplemented
without fussing. That is, once we depend on astropy>=6.0...

See https://github.com/astropy/astropy/pull/12888 for more details.