SPAZA-299 Ignore sep for decimal angle strings
The `Angle.to_string` code changed in astropy/astropy#13933 so that
it is now an error to specify a sexagesimal separator (`sep`) when
a decimal representation is requested (`decimal=True`). This was
released in Astropy 5.2 and eventually backported to 5.0.

Based on the error message I suspect that this behaviour is
coincidental. The commit message talks about allowing a unit for
the "generic" string format and the exception mentions the units too.

Either way, the katpoint `angle_to_string` function is meant to be
more tolerant / bumbling, so simply ignore the separator when decimal
is True. Update the docstring to mention this and other quirks.