SPAZA-299 Print out-of-bound coordinates better
When a coordinate is outside the expected tolerance in the unit tests,
print it in the format in which it is specified in the tests to make
it easy to copy and paste the correct value back into the tests.
Also print the separation in the same unit as the tolerance.