SPAZA-299 Fix unit tests on Astropy 5.3
Astropy 5.3 introduced a new IERS_B table based on the ITRF2020 model
of the Earth - see astropy/astropy#14382 for more details. The model
came into effect in the IERS tables on 14 Feb 2023 and made it into
Astropy 5.3 from there. The new frame definition causes a jump in
UT1-UTC of about 0.05 ms, or an Earth rotation of about 1 mas,
affecting topocentric coordinates like azel. Pick the Astropy 5.3
value as azel reference wherever it failed, and adjust the tolerance
accordingly. The tests still pass on all minor versions of Astropy
from 4.1 to 5.3.