SPAZA-299 Improve unit test tolerances
Specify RA seconds with five decimal places to get an accuracy of the
same order of magnitude as for the Dec in degrees. This is based on
values obtained with Astropy 5.3.1.

Fix the helper routine to print out RA in hourangle instead of degrees.

Fix the "accurate" tolerance to 0.1 mas for both radec and azel, since
this represents the combination of RA and Dec accuracies in a
quadrature sum. Improve the tolerances in the Great Conjunction
example, which seems to hold for multiple Astropy versions.

The tests pass on all Astropy versions >= 4.1.