SPAZA-299 Use Astropy 5.3 as reference in test
Even the StationaryBody version of `test_compute` needs to move to
Astropy 5.3 to be accurate, since the (ra, dec) is slightly different
due to the shift in the Earth model to ITRF2020.

Rename `min_astropy_ver` to be more descriptive. It's the minimum
Astropy version to qualify for the highest accuracy, but it is better
described as the reference version.