SPAZA-299 Add PyPI as primary package source
This follows the suggestion of this recent Poetry warning:

  In a future version of Poetry, PyPI will be disabled automatically
  if at least one custom source is configured with another priority
  than 'explicit'.

Downgrade our SKAO source to "supplemental" (the old "secondary"),
since it is currently only useful for the documentation theme.

While we are at it, shuffle the config file around a bit. And change
the dev dependencies to a proper group (another futureproofing step).