SPAZA-299 More docs-build massaging
We need to install katpoint to get the version number. We therefore
also need its dependencies. This means that we have to overcome the
default settings of `--no-root --only docs`. Override this in the
docs-pre-build make target.

Redo the readthedocs YAML according to ska-cookiecutter-pypackage
but with our versions and setup.