SPAZA-299 Get version from git tags with Poetry
This ends the brave journey of katversion for katpoint ;-D It is too
hard to reconcile with Poetry, and there is an alternative.

Use the poetry-dynamic-versioning plugin instead, based in turn on
Dunamai. The process is as follows:

  - Tag the git commit that you want to release with a version number
    (PEP440 for now, SemVer in the future). Commits after that will
    automatically be treated as development versions.
  - During `poetry build` and `poetry install` the dummy package
    version of "0.0.0" is replaced by the actual version.
  - The version info is also written to `src/katpoint/_version.py` from
    where it is imported and exposed as `katpoint.__version__`. This
    file is ignored otherwise and not committed to the repository.

In order for this to work, set `GIT_DEPTH = 0` to avoid shallow Git
repositories in the GitLab CI pipeline. This has no speed impact
because the repo is small. Refresh the poetry installation in various
CI stages to ensure that the dynamic version takes effect.

It would be nice to support "dirty" version numbers but unfortunately
the CI pipeline tweaks the version in pyproject.toml behind our backs
(if a "non_tag_pyproject" build), which renders those builds "dirty".