SPAZA-299 Upgrade to the latest Poetry 1.7.1
Old Poetry is painful. The `CI_POETRY_VERSION` currently applies to the
following jobs:

  - `python-lint`
  - `python-test`
  - `docs-build`
  - `docs-pages`

The `python-build` job doesn't really need it as the PEP 517 machinery
will pull in an appropriate Poetry.

It would be nice if `python-gemnasium-scan` obeyed it too, but not yet.
(I've basically given up on scan after many hours of sweat and tears.)

Rebuild the lock file and document the package versions used to
generate it. This includes the latest poetry-dynamic-versioning as well.