SPAZA-299 Fix docs-build by enforcing lock file
Use `poetry install` instead of `poetry add` to honour the lock file.
The old code pulled in newer versions of poetry-dynamic-versioning and
its dependencies, which somehow broke the pipeline. The lock file
already specifies "poetry-dynamic-versioning[plugin]" so no need to
add this explicitly.

We need to call `poetry install` twice so that the version is correct
by the time `--with docs` is invoked.

In future we could consider removing the need to import katpoint in
`conf.py` and pull the version from `pyproject.toml` instead -- see e.g.
ska-sdp-realtime-receive-core's conf.py.