SPAZA-300 Rename RefractionCorrection class
It is not clear which way the correction goes (refracted -> unrefracted
or vice versa), making this a source of confusion. The code also does
both directions, so it's more than just the correction. Once the `apply`
and `reverse` methods are also renamed, all should be clear :-)

The new name (TroposphericRefraction) mirrors that of TroposphericDelay
and highlights its spot in the module hierarchy.