SPAZA-300 Change TroposphericRefraction API
Rename apply -> refract and reverse -> unrefract to be clearer about
the direction in which corrections are applied. "Refract" switches
on the atmosphere and typically increases the elevation angle.
"Unrefract" switches off the atmosphere, thereby decreasing the
elevation angle.

Use the same factory mechanism as TroposphericDelay to instantiate
the selected model, and also call the model name `model_id`. The idea
is to make the object read-only, so that `model_id` unambiguously
represents the actual algorithm in use.

The actual model now becomes a class with two class methods applying
and removing refraction, respectively, instead of a function doing
one direction only. The original RefractionCorrection had a fancy
"refract" step and a generic function inversion as "unrefract". However,
it turns out that most refraction models have a fancy "unrefract" step
instead, which kills the dream of only specialising the "refract"
direction.

Keep the original katpoint refraction algorithm for starters, now called
"HaystackRefraction" in deference to its origin in the MIT Haystack
Pointing System back in the 1960's. Calling it "FieldSystemRefraction"
could be misleading since the FS uses Lanyi's algorithm these days and
also has the DSN refraction model ("sbend") floating around.