SPAZA-300 Use Astropy units in refraction
Change weather parameters to (pressure, temperature, relative_humidity)
to match TroposphericDelay and Astropy. I left the basic Haystack
algorithm as intact as possible, since the units of those coefficients
are wacky anyway. Change the tolerances to more human-readable versions
though. The elevation is converted to an Angle object to allow easy
rad/deg conversion and flexible construction. The input parameter is
therefore of type "Angle-like".

Add a unit test to compare the actual values against those of katpoint
0.10, to ensure that the code does not regress. Previously we only
checked the roundtripping between refract and unrefract, not that
refract is actually correct :-)