SPAZA-300 Redo tropospheric classes as dataclasses
It is a good idea to make TroposphericRefraction and TroposphericDelay
"read-only" in some sense because their data members mostly revolve
around the identity of the underlying model, which should not change
after instantiation.

On the other hand, it is hard to make a generic Python object truly
immutable, as there are various ways to modify it behind the scenes.
A reasonable approach is to suppress the __setattr__ method, which was
done up to now.

A more modern approach is to use frozen dataclasses, available in the
standard library since Python 3.7. This removes a lot of boilerplate
methods in the case of TroposphericRefraction and makes the intent
clearer. Since CAM still needs Python 3.6, install the backported
package in that case.

In the process, add a `_RefractionModel` base class, which will be
useful once ERFA refraction lands. This simplifies the class lookup
by name as well.

Add various tests of the object properties, which caught a string
formatting bug in the error message of TroposphericDelay :-)