SPAZA-300 Clarify the type of _model attribute
The `TroposphericRefraction._model` attribute had the type hint
`_RefractionModel`, which implies that it is an instance of this class.
Actually it is the relevant subclass itself, because the underlying
classes only have class methods (refract and unrefract) and therefore
do not need instantiation. Fix the description with a TypeVar (one
step closer to mypy :-) ).