SPAZA-300 TroposphericRefraction operates on AltAz
Since it is somewhat inconvenient to split AltAz objects into (az, el),
transform the elevation angles and assemble it back into AltAz, let
`TroposphericRefraction` take care of it. It therefore becomes a
direction transform instead of an elevation transform.

Clip the transformed elevation angles to fit into the range [-90, 90]
degrees, otherwise Latitude complains bitterly. This should be safe...
If refracting, the refraction offset at zenith will be tiny but will go
over the top. It can be safely ignored. If unrefracting, elevations
will decrease, which is potentially an issue at -90 degrees, where the
offset will be large but simultaneously unphysical what with the Earth
in the way and all :-)