SPAZA-300 New ERFA-based refraction model
Expose the refraction routines used internally by Astropy to transform
to and from the observed frame, which in turn comes from ERFA.

The `unrefract` step implements a model dZ = A tan Z + B tan^3 Z, where
Z is the zenith angle / distance and constants A and B (in radians) are
derived from the weather parameters by the `erfa.refco` / eraRefco
routine. This routine in turn traces its lineage to SOFA and the
iauRefco routine, and SLALIB and the sla_REFCOQ routine before that.
The implementation is a translation of the relevant bit of code found
in `erfa.atoiq` / eraAtoiq.

The `refract` step is very similar to the SLALIB sla_REFV routine,
which inverts the tan(z) model with a single Newton-Raphson iteration.
This approach first popped up in SOFA release 10 in 2013, which
introduced the various high-level astrometry routines. This
implementation follows the code in `erfa.atioq` / eraAtioq, which in
turn derives from SOFA's iauAtioq. Break out the "eraRefv" routine as
it might be useful in general.

Of course, after writing all this, I discovered that Astropy 5.2
introduced `add_refraction` and `remove_refraction` routines as part
of topocentric ITRS frames. Since we can't jump to Python 3.8 yet,
we are forced to continue with our homegrown (and equivalent) approach
for now :-)

Test our versions of the refraction routines against Astropy's
transforms as a doublecheck. They should be identical within numerical
precision.

I considered calling the class ERFARefraction in deference to PEP 8 but
Astropy speaks of ErfaWarnings and ErfaErrors so I followed their lead.