SPAZA-300 Remove units from low-level models again
It's too slow... Especially `@u.quantity_input`.

The lower-level _RefractionModels operate on elevation angles and are
fully unitised. The problem is that the units will be checked at the
TroposphericRefraction and _RefractionModel levels, doubling the time
it takes for no good reason.

For example, a call to `TroposphericRefraction.unrefract` took about
500 us on my laptop and `ErfaRefration.unrefract` made up 250 us of
that, while the underlying ERFA routines only took about 2 us.

Revert the routines operating on elevation angles to use straight floats
and/or arrays with explicit units, like it was before. Just standardise
on fractional relative humidity instead of percentages. All unit checks
now happen on the TroposphericRefraction level, which is the main API
for this functionality.

I considered making ErfaRefraction and HaystackRefraction private (i.e.
_ErfaRefraction and _HaystackRefraction) to discourage its use. On the
other hand, it gives access to fast routines which could be useful for
e.g. benchmarking and is already somewhat hidden in the submodule
katpoint.troposphere.refraction, therefore keep it as is.