SPAZA-300 Fix Poetry lock content hash
Poetry is nice for applications running very specific versions but
problematic for libraries that have to support a wide range of versions.
Katpoint is one of the latter. It still has to run on Python 3.6 but
the SKAO CI environment runs on Python 3.8+.

My solution has been to fix the lock file to the CI environment,
representing a reasonably typical use case, while the pyproject.toml
file supports the full gamut of versions (even if there might be
conflicts between some of the dependencies in extreme cases). To achieve
this, I effectively use a different version of the pyproject.toml to
make the lock file (see comments in pyproject.toml).

It turns out that Poetry stores a hash of the relevant pyproject.toml
section in the lock file and wants them to match up. In the past a
mismatch resulted in a warning, but since Poetry 1.8.0 this has turned
into an error (GitHub python-poetry/poetry#8737).

We can't do that. That messes up our plan.

My workaround is to generate the hash of the actual pyproject.toml (not
the one used to construct the lock file) and store that in the lock file
instead. Get the hash by running Python within a `poetry shell` and
doing

```Python
from poetry.console.application import Application

app = Application()
print(app.poetry.locker._get_content_hash())
```

This makes `poetry check --lock` pass. A thing of beauty :-)