SPAZA-123 Benchmark tests and adjust accuracies
Standardise on N=100000 when determining the tolerances for closure
tests. Because we'll run the tests routinely with N=100 anyway (which
does much better, typically 12 decimals), there is no point to make
these tolerances too loose. They seem to have been derived based on
N=1000000 in the past, and then frequently a single failure results in
a lower threshold.

A quick benchmark using `pytest --durations=0` and N=1000000 on my
laptop reveals:

- test_random_closure (3 katpoint calls) takes 0.3 s
- test_aips_compatibility (2 katpoint and 2 AIPS calls) takes 1.1 s
- test_wcs_compatibility (2 katpoint and 2 wcs calls) takes 6.7 s

So WCSLIB via the Astropy layers is about 8x slower than FITS, which in
turn is 4x slower than the katpoint projection routines, when used on
vector data. This is not too surprising, since we are calling both AIPS
and `astropy.wcs` one coordinate at a time. Granted, all of them are
still fast - even `astropy.wcs` does 1000 projections in 3 ms.

The accuracy is similar, with the same tolerances used when comparing
katpoint to WCSLIB and AIPS.