SPAZA-123 Remove AIPS projection test machinery
Now that we have Astropy WCSLIB as a benchmark, there is no more need
for the somewhat clumsy AIPS download + compile script and the unit
tests that use the resulting extension module. These tests were
typically skipped in the CI pipeline anyway, so now there are fewer
skipped tests, which is a good thing.