SPAZA-228 Fix parallactic angle
Instead of a convoluted procedure involving apparent radec, do a direct
position angle measurement between ICRS North and zenith. This seems
like the most useful formulation of parallactic angle.

Also replace the global North celestial pole with a local version,
obtained by moving slightly North of the target. This handles the fact
that aberration introduces a non-rigid distortion of the celestial
sphere that does not preserve great circles. The global parallactic
angle therefore differs from the local one, and the latter is preferred.
The small step to the North matches the value introduced by @bmerry in
https://github.com/ska-sa/katsdpimager/pull/17/ (about 2 arcseconds,
much smaller than the 0.03 radian step in `Target.uvw_basis`).

It is interesting to note that the position angle of the NCP in (az, el)
equals the position angle of zenith in ICRS. Even though the direction
of the angle is different, the angular value is the same because ICRS
and AltAz have different handedness when it comes to East vs North.
Although parallactic angle is defined as the celestial position angle
of zenith, we obtain it as the horizontal position angle of the NCP
instead because it can potentially be sped up in future in the case of
multiple antennas.