SPAZA-302 Check that pip install works
Since the CI pipeline is so Poetic, it missed that the katpoint wheel
is completely broken (see the very first issue #1 :-) ). Add an explicit
`pip install` step at the end of the python-test job so that it doesn't
interfere with the normal tests. We just want to be sure that a basic
import works after `pip install`.