SPAZA-303 Add bits of 0.10 to changelog
The original v1.0 branch started from the 0.9 release. The first 1.0a1
release on GitLab pretty much coincided with the 0.10 release of new
functionality on GitHub. I then ported the relevant bits of 0.10 across
to the GitLab master so that it could be feature complete (SPAZA-297).
This backport only happened after 1.0a2 and will now be included in
the 1.0a3 release.

This commit updates the changelog to reflect the 0.10 release when it
happened on GitHub, but restricts the listed items to those that made it
into 1.0. So a bit of a kludge. :-)

I prefer this way as opposed to including the items under the main 1.0a3
heading, because it still refers to GitHub pull requests, for one.