SPAZA-303 Fix copyright year to match commit dates
This is the "Merrian" style (sorry, Bruce!), which avoids grand
"bump copyright year" commits in favour of updating the copyright
of files you edit as part of the edit. This allows a quick check
of when a file was last changed and avoids spurious commits.

I did not include years in the copyright list if the copyright notice
was the only thing that changed in the file that year. If the list
became too cumbersome, I collapsed some of the year ranges.

In addition, fix the following:

- scripts/psrcat_sources.py had no copyright notice yet.
- src/katpoint/__init__.py still referred to SKA instead of SARAO.