SPAZA-303 Upgrade to Poetry>=1.8.2 and redo lock
The latest ska-cicd-k8s-tools-build-deploy:0.13.1 CI image (for the
build and deploy steps :-)) uses Poetry 1.8.2 while we were downgrading
it to 1.7.1. This fixes some inexplicable errors that crept into the
`docs-build` job.

(Unfortunately the ska-tango-images-pytango-builder:9.3.35 CI image
used by `python-lint` and `python-test` remains on Poetry 1.3.2, but
there the upgrade to 1.8.2 goes smoothly.)

Also make `twine` an explicit dev dependency instead of installing it
after the fact. Fix the lock hash as described in commit a02062f.