SPAZA-123 WCS projection needs 1-D arrays
While the ideal is that all projection routines should support arbitrary
(but commensurate) array shapes, this is not yet true of the WCS-based
routines. These are forced to use non-empty 1-D arrays, since those
are either iterated over to produce scalars, or passed in as-is to the
underlying `astropy.wcs` routines. Note this in the docstrings.